# opt: improve performance for tables with many columns and indexes